Object Oriented Development coursework (2nd year).
Using javafx library for creating GUI for chosing modules for a final year at the university. The job was to develop View and Controlled parts of MVC. The model was provided by the module leader. If I had to do it again, I'd avoid the spaghettification of the Controller file by abstracting functionality to separate files. final grade was (86/100). The program fulfilled all coursework requirements